// src/models/content-types/band.ts
/** 
* This file has been auto-generated by '@kontent-ai/model-generator@8.1.2'.
* 
* (c) Kontent.ai
*  
* -------------------------------------------------------------------------------
* 
* Project: Concerts
* Environment: Production
* Id: c1ad4901-f748-000b-2b83-2c4fa51e2983
* 
* -------------------------------------------------------------------------------
**/

import type { Elements } from "@kontent-ai/delivery-sdk";
import type { CoreContentType } from "../system/index.ts";
import type { Genre } from "../taxonomies/index.ts";
import type { Person, PlaylistSong } from "./index.ts";

/**
 * Band
 *
 * Id: d23949c5-622d-4bd7-a752-dcb1f991e0bf
 * Codename: band
 */
export type Band = CoreContentType<
  BandElementCodenames,
  {
    /**
     * Band Name
     *
     * Type: text
     * Required: true
     * Codename: band_name
     * Id: 588f6253-029e-46b9-8ef2-673f8dd74f91
     * Guidelines: Name of the band/performing artist
     */
    readonly band_name: Elements.TextElement;
    /**
     * Band Logo
     *
     * Type: asset
     * Required: true
     * Codename: band_logo
     * Id: cb31087f-f15a-48f5-9e4a-ceb72bf54bbf
     * Guidelines: Utilize the most current logo image from band/artist's press kit.
     */
    readonly band_logo: Elements.AssetsElement;
    /**
     * Promotional Image
     *
     * Type: asset
     * Required: false
     * Codename: promotional_image
     * Id: 8cc60164-e242-45aa-8aca-5e4b0f89f3c7
     * Guidelines: Utilize the promotional image provided from the band/artist's press kit.
     */
    readonly promotional_image: Elements.AssetsElement;
    /**
     * Band Bio
     *
     * Type: rich_text
     * Required: true
     * Codename: band_bio
     * Id: 96f6c14f-b298-4f48-9f21-05505660b46b
     * Guidelines: Use the bio provided in the band/artist's press kit. If none is provided, draft a promotional bio around three paragraphs in length.
     */
    readonly band_bio: Elements.RichTextElement<CoreContentType>;
    /**
     * Band Member
     *
     * Type: modular_content
     * Required: true
     * Codename: band_member
     * Id: 8c74a2a4-1e4d-48fc-a431-0cc4b38419c5
     * Guidelines: Include the current members of the band.
     */
    readonly band_member: Elements.LinkedItemsElement<Person>;
    /**
     * Social Media
     *
     * Type: rich_text
     * Required: false
     * Codename: social_media
     * Id: 0d33cbb4-ab04-423a-a8b9-532602e8c80a
     * Guidelines: Insert band/artist's social media links.
     */
    readonly social_media: Elements.RichTextElement<CoreContentType>;
    /**
     * Band Playlist
     *
     * Type: modular_content
     * Required: true
     * Codename: band_playlist
     * Id: cf85fe66-1563-459e-9d05-6852e84113b2
     * Guidelines: Include three to five of the band/artist's songs that are most popular or otherwise representational of their body of work.
     */
    readonly band_playlist: Elements.LinkedItemsElement<PlaylistSong>;
    /**
     * Genre
     *
     * Type: taxonomy
     * Required: true
     * Codename: band_genre
     * Id: e2228fdc-3794-4ab0-b4da-8ec19feefaf3
     * Guidelines: Indicate music genres most relevant to the band/artist's work (multiple genres can be selected)
     */
    readonly band_genre: Elements.TaxonomyElement<Genre, "band_genre">;
    /**
     * slug
     *
     * Type: url_slug
     * Required: false
     * Codename: slug
     * Id: 05302667-ee40-4042-99d6-7fbf451d21ef
     */
    readonly slug: Elements.UrlSlugElement;
  },
  "band"
>;

/**
 * Type representing all available element codenames for Band
 */
export type BandElementCodenames =
  | "band_name"
  | "band_logo"
  | "promotional_image"
  | "band_bio"
  | "band_member"
  | "social_media"
  | "band_playlist"
  | "band_genre"
  | "slug";

/**
 * Type guard for Band
 *
 * Id: d23949c5-622d-4bd7-a752-dcb1f991e0bf
 * Codename: band
 */
export function isBand(item: CoreContentType | undefined | null): item is Band {
  return item?.system?.type === "band";
}