// src/models/content-types/venue.ts
/** 
* This file has been auto-generated by '@kontent-ai/model-generator@8.1.2'.
* 
* (c) Kontent.ai
*  
* -------------------------------------------------------------------------------
* 
* Project: Concerts
* Environment: Production
* Id: c1ad4901-f748-000b-2b83-2c4fa51e2983
* 
* -------------------------------------------------------------------------------
**/

import type { Elements } from "@kontent-ai/delivery-sdk";
import type { CoreContentType } from "../system/index.ts";
import type { Location } from "./index.ts";

/**
 * Venue
 *
 * Id: fb0081ad-34a0-4c69-b262-7cec3bad6b4b
 * Codename: venue
 */
export type Venue = CoreContentType<
  VenueElementCodenames,
  {
    /**
     * Venue Name
     *
     * Type: text
     * Required: true
     * Codename: venue_name
     * Id: dfb65a08-05dc-4415-9c44-07ff8e4f0118
     * Guidelines: Name of the venue.
     */
    readonly venue_name: Elements.TextElement;
    /**
     * Hero Image
     *
     * Type: asset
     * Required: false
     * Codename: hero_image
     * Id: 86b003e1-5e49-4f93-aa1b-37af71880aab
     * Guidelines: Promotional image of the venue.
     */
    readonly hero_image: Elements.AssetsElement;
    /**
     * Location
     *
     * Type: modular_content
     * Required: true
     * Codename: location
     * Id: 6780fee5-6e72-4f75-adf5-a6ab08b45341
     * Guidelines: Full address and phone number of the venue.
     */
    readonly location: Elements.LinkedItemsElement<Location>;
    /**
     * Venue Slug
     *
     * Type: url_slug
     * Required: true
     * Codename: venue_slug
     * Id: d98b9243-e16f-4c54-a91b-49fcdb5c4dc4
     */
    readonly venue_slug: Elements.UrlSlugElement;
  },
  "venue"
>;

/**
 * Type representing all available element codenames for Venue
 */
export type VenueElementCodenames =
  | "venue_name"
  | "hero_image"
  | "location"
  | "venue_slug";

/**
 * Type guard for Venue
 *
 * Id: fb0081ad-34a0-4c69-b262-7cec3bad6b4b
 * Codename: venue
 */
export function isVenue(
  item: CoreContentType | undefined | null,
): item is Venue {
  return item?.system?.type === "venue";
}