// src/models/content-types/concertReview.ts
/** 
* This file has been auto-generated by '@kontent-ai/model-generator@8.1.2'.
* 
* (c) Kontent.ai
*  
* -------------------------------------------------------------------------------
* 
* Project: Concerts
* Environment: Production
* Id: c1ad4901-f748-000b-2b83-2c4fa51e2983
* 
* -------------------------------------------------------------------------------
**/

import type { Elements } from "@kontent-ai/delivery-sdk";
import type { Concert, Person } from "./index.ts";
import type { CoreContentType } from "../system/index.ts";

/**
 * Concert Review
 *
 * Id: 531d8f36-e963-42e1-a9a6-5b92ba91bc24
 * Codename: concert_review
 */
export type ConcertReview = CoreContentType<
  ConcertReviewElementCodenames,
  {
    /**
     * Concert
     *
     * Type: modular_content
     * Required: true
     * Codename: concert
     * Id: f717fc82-6a04-4a08-8b0f-037c57d59ab6
     * Guidelines: Link the concert being reviewed
     */
    readonly concert: Elements.LinkedItemsElement<Concert>;
    /**
     * Headline
     *
     * Type: text
     * Required: true
     * Codename: headline
     * Id: 72d7db93-dafa-4dcd-bb3a-8ae51d8c4ace
     * Guidelines: Write an Interesting headline for the concert review.
     */
    readonly headline: Elements.TextElement;
    /**
     * Author
     *
     * Type: modular_content
     * Required: true
     * Codename: author
     * Id: 185df368-f182-4b54-8f38-4bc9798ffed8
     * Guidelines: Link the author of the review
     */
    readonly author: Elements.LinkedItemsElement<Person>;
    /**
     * Review Body
     *
     * Type: rich_text
     * Required: true
     * Codename: review_body
     * Id: 698ed1ca-19ce-47eb-9020-c8d7ba5ee93c
     * Guidelines: Write the body of the review here
     */
    readonly review_body: Elements.RichTextElement<CoreContentType>;
    /**
     * Untitled url slug
     *
     * Type: url_slug
     * Required: true
     * Codename: untitled_url_slug
     * Id: 6266e3d2-ccd3-47e6-ad78-c0d23f95150a
     */
    readonly untitled_url_slug: Elements.UrlSlugElement;
  },
  "concert_review"
>;

/**
 * Type representing all available element codenames for Concert Review
 */
export type ConcertReviewElementCodenames =
  | "concert"
  | "headline"
  | "author"
  | "review_body"
  | "untitled_url_slug";

/**
 * Type guard for Concert Review
 *
 * Id: 531d8f36-e963-42e1-a9a6-5b92ba91bc24
 * Codename: concert_review
 */
export function isConcertReview(
  item: CoreContentType | undefined | null,
): item is ConcertReview {
  return item?.system?.type === "concert_review";
}